Adding connectedToTheDom
I found another use case that was letting through backspace-related browser history state changes: If a backspace event results in removing the element being typed in from the dom, the default action ends up being a history change. So in the case a field has been removed from the dom, the default must also be prevented.

I also changed the validInputTypes to an object for performance and readability, and made a number of other minor readability improvements.